requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.19
torch==2.7.1
torchvision==0.22.1
transformers==4.52.4
pillow==11.2.1
h11~=0.16.0
pip~=25.1.1
wheel~=0.45.1
filelock~=3.18.0
PyYAML~=6.0.2
anyio~=3.7.1
sniffio~=1.3.1
typing_extensions~=4.14.0
click~=8.2.1
uvloop~=0.21.0
starlette~=0.27.0
pydantic_core~=2.14.1
requests~=2.32.3
websockets~=15.0.1
httptools~=0.6.4
watchfiles~=1.0.5
annotated-types~=0.7.0
Jinja2~=3.1.6
pandas~=2.3.0
numpy~=2.3.0